requirements.txt
# Kenya Clinical Reasoning Challenge - Core Dependencies
# Production-grade medical AI system

# Core ML/AI frameworks
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
torch>=2.0.0
transformers>=4.30.0

# ML Evaluation and Training
rouge-score>=0.1.2
datasets>=2.8.0
accelerate>=0.15.0

# Medical NLP and text processing
spacy>=3.6.0
medspacy>=1.0.0
nltk>=3.8.0
textstat>=0.7.0

# Data visualization and analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0
loguru>=0.7.0

# Evaluation metrics
bert-score>=0.3.13
evaluate>=0.4.0

# Medical knowledge integration
requests>=2.31.0
beautifulsoup4>=4.12.0

# Development and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
jupyter>=1.0.0

# Optional GPU acceleration
# torch-audio>=2.0.0  # Uncomment if needed
# torch-vision>=0.15.0  # Uncomment if needed
